reflection: clang format

+++ ola_front_protocol_core.hpp
@@ -14,43 +14,44 @@ inline constexpr const char* default_port()
 }
 using ProtocolTypeIndexT = std::pair<uint8_t, uint16_t>;
 
-namespace core{
+namespace core {
 
-constexpr size_t protocol_id = 0;
-    
-struct Version{
-    static constexpr uint32_t version = 1;
-    static constexpr uint32_t auth_request = 1;
+constexpr uint8_t protocol_id = 0;
+
+struct Version {
+    static constexpr uint32_t version       = 1;
+    static constexpr uint32_t auth_request  = 1;
     static constexpr uint32_t auth_response = 1;
     static constexpr uint32_t init_response = 1;
-    static constexpr uint32_t response = 1;
-    
-    uint32_t version_ = version;
-    uint32_t auth_request_ = auth_request;
+    static constexpr uint32_t response      = 1;
+
+    uint32_t version_       = version;
+    uint32_t auth_request_  = auth_request;
     uint32_t auth_response_ = auth_response;
     uint32_t init_response_ = init_response;
-    uint32_t response_ = response;
-    
-    void clear(){
+    uint32_t response_      = response;
+
+    void clear()
+    {
         auth_request_ = auth_response_ = response_ = -1;
     }
-    
-    bool operator<=(const Version &_rthat)const{
-        return  version_ <= _rthat.version_ && auth_request <= _rthat.auth_request_ &&
-                auth_response_ <= _rthat.auth_response_ && init_response_ <= _rthat.init_response_ &&
-                response_ <= _rthat.response_;
+
+    bool operator<=(const Version& _rthat) const
+    {
+        return version_ <= _rthat.version_ && auth_request <= _rthat.auth_request_ && auth_response_ <= _rthat.auth_response_ && init_response_ <= _rthat.init_response_ && response_ <= _rthat.response_;
     }
-    
-    SOLID_REFLECT_V1(_s, _rthis, _rctx){
+
+    SOLID_REFLECT_V1(_s, _rthis, _rctx)
+    {
         _s.add(_rthis.version_, _rctx, 1, "version");
         _s.add([&_rthis](Reflector& _s, Context& _rctx) {
-            if constexpr (!Reflector::is_const_reflector){
-                if(_rthis.version > Version::version){
+            if constexpr (!Reflector::is_const_reflector) {
+                if (_rthis.version > Version::version) {
                     _rthis.clear();
                     return;
                 }
             }
-            if(_rthis.version_ == version){
+            if (_rthis.version_ == version) {
                 _s.add(_rthis.auth_request_, _rctx, 3, "auth_request");
                 _s.add(_rthis.auth_response_, _rctx, 4, "auth_response");
                 _s.add(_rthis.init_response_, _rctx, 5, "init_response");
@@ -78,7 +79,7 @@ struct AuthRequest : solid::frame::mprpc::Message {
 };
 
 struct AuthResponse : solid::frame::mprpc::Message {
-    uint32_t    error_   = -1;
+    uint32_t    error_ = -1;
     std::string message_;
 
     AuthResponse() {}
@@ -96,9 +97,9 @@ struct AuthResponse : solid::frame::mprpc::Message {
 };
 
 struct InitResponse : solid::frame::mprpc::Message {
-    uint32_t    error_   = -1;
+    uint32_t    error_ = -1;
     std::string message_;
-    
+
     InitResponse() {}
 
     InitResponse(
@@ -109,14 +110,14 @@ struct InitResponse : solid::frame::mprpc::Message {
 
     SOLID_REFLECT_V1(_s, _rthis, _rctx)
     {
-        if(_rctx.anyTuple().template getIf<Version>()->init_response_ == Version::init_response){
+        if (_rctx.anyTuple().template getIf<Version>()->init_response_ == Version::init_response) {
             _s.add(_rthis.error_, _rctx, 1, "error").add(_rthis.message_, _rctx, 2, "message");
         }
     }
 };
 
 struct Response : solid::frame::mprpc::Message {
-    uint32_t    error_   = -1;
+    uint32_t    error_ = -1;
     std::string message_;
 
     Response() {}
